backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-dotenv==1.0.0
pyyaml==6.0.1
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
